genderclassification/requirements.txt
Pillow==9.5.0
streamlit==1.22.0
torch==2.0.0
torchvision==0.15.1
